+++ src/components/general-widgets/FeaturedRoomTypes/constants.js
@@ -3,7 +3,7 @@ const PLACEHOLDER_DATA = {
   guestsNumber: 0,
   imageUrl: '',
   locationName: '',
-  nightPrice: '',
+  pricePerPeriod: '',
   roomTypeName: '',
   roomTypeUrl: '',
 };