+++ src/utils/get-night-price-markup/index.js
@@ -1,1 +1,0 @@
-export { getNightPriceMarkup } from './getNightPriceMarkup';